Move Identifier to own file
